Fix
Former-commit-id: 3ef0950e81c87435033121596f4cae2c732b90a5

+++ backend/src/main/java/com/ssafy/moeutto/domain/aiRecOutfit/dto/request/AiRecOutfitCombineRequestDto.java
@@ -18,13 +18,13 @@ public class AiRecOutfitCombineRequestDto {
     private Integer pty; // 강수 상태
 
     @NotNull
-    private Double tmn; // 최저 기온
+    private Float tmn; // 최저 기온
 
     @NotNull
-    private Double tmx; // 최고 기온
+    private Float tmx; // 최고 기온
 
     @NotNull
-    private Double wsd; // 풍속
+    private Float wsd; // 풍속
 
     @NotNull
     private Date date; // 날짜 정보
